+++ src/helpers/functions/Role.js
@@ -1,5 +1,7 @@
-const { roles, sideNames } = require("../constants/Role");
+const { roles, sideNames, groupNames } = require("../constants/Role");
 
-exports.getPlayerRoles = () => JSON.parse(JSON.stringify(roles));
+exports.getRoles = () => JSON.parse(JSON.stringify(roles));
 
 exports.getSideNames = () => JSON.parse(JSON.stringify(sideNames));
+
+exports.getGroupNames = () => JSON.parse(JSON.stringify(groupNames));