+++ tests/e2e/game/full-game.test.js
@@ -295,7 +295,7 @@ describe("B - Full game of 18 players with all roles", () => {
             .end((err, res) => {
                 expect(res).to.have.status(200);
                 game = res.body;
-                expect(game.players[16].role.group).to.deep.equals("werewolves");
+                expect(game.players[16].side).to.deep.equals("werewolves");
                 done();
             });
     });
@@ -2158,7 +2158,7 @@ describe("B - Full game of 18 players with all roles", () => {
             });
     });
     it("🐒 Wild child changed his side to `werewolves` because his model (the werewolf) just died", done => {
-        expect(game.players[5].role.group).to.equals("werewolves");
+        expect(game.players[5].side).to.equals("werewolves");
         done();
     });
     it("🎲 Game is waiting for 'sheriff' to 'delegate'", done => {